pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.3.1",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "groq>=1.0.0",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20231228",
    "pillow>=10.0.0",
    "pinecone>=8.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.0",
    "pytesseract>=0.3.10",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.40.0",
]
